packages/tools-git/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yagura-tools-git"
version = "0.1.0"
description = "Git operations (status, log, diff, commit, push, PR) for the Yagura agent framework."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Kouta", email = "sk-kouta-dev@users.noreply.github.com" }
]
license-files = ["LICENSE"]
keywords = ["ai", "agent", "yagura", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "yagura-agent>=0.1.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
github = ["PyGithub>=2.0"]
gitlab = ["python-gitlab>=4.0"]
dev = ["pytest", "pytest-asyncio"]

[project.urls]
Homepage = "https://github.com/sk-kouta-dev/yagura"
Documentation = "https://github.com/sk-kouta-dev/yagura"
Repository = "https://github.com/sk-kouta-dev/yagura"

[tool.hatch.build.targets.wheel]
packages = ["yagura_tools"]
